Hernoem package van nl.rijksoverheid.moz.berichtenmagazijn naar nl.rijksoverheid.moz.fbs.berichtenmagazijn
