--- services/berichtenmagazijn/src/test/kotlin/nl/rijksoverheid/moz/berichtenmagazijn/aanlever/BerichtOpslagServiceTest.kt
@@ -1,43 +1,0 @@
-package nl.rijksoverheid.moz.berichtenmagazijn.aanlever
-
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.slot
-import io.mockk.verify
-import nl.rijksoverheid.moz.berichtenmagazijn.opslag.Bericht
-import nl.rijksoverheid.moz.berichtenmagazijn.opslag.BerichtRepository
-import nl.rijksoverheid.moz.berichtenmagazijn.opslag.IdentificatienummerType
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNotNull
-import org.junit.jupiter.api.Test
-
-class BerichtOpslagServiceTest {
-
-    private val repository = mockk<BerichtRepository>(relaxed = true)
-    private val service = BerichtOpslagService(repository)
-
-    @Test
-    fun `opslaanBericht roept repository opslaan aan en retourneert het domeinobject`() {
-        val berichtSlot = slot<Bericht>()
-        every { repository.save(capture(berichtSlot)) } answers { }
-
-        val bericht = service.opslaanBericht(
-            afzender = "00000001003214345000",
-            ontvangerType = IdentificatienummerType.BSN,
-            ontvangerWaarde = "999993653",
-            onderwerp = "Voorlopige aanslag 2026",
-            inhoud = "Hierbij ontvangt u...",
-        )
-
-        assertNotNull(bericht.berichtId)
-        assertNotNull(bericht.tijdstipOntvangst)
-        assertEquals("00000001003214345000", bericht.afzender.waarde)
-        assertEquals(IdentificatienummerType.BSN, bericht.ontvanger.type)
-        assertEquals("999993653", bericht.ontvanger.waarde)
-        assertEquals("Voorlopige aanslag 2026", bericht.onderwerp)
-        assertEquals("Hierbij ontvangt u...", bericht.inhoud)
-
-        verify { repository.save(any<Bericht>()) }
-        assertEquals(bericht, berichtSlot.captured)
-    }
-}
